migrate to netcore8.0 and mysql9.0

+++ SimpleModulus/Network/Packet.cs
@@ -424,7 +424,8 @@ namespace WebZen.Network
                 }
                 else
                 {
-                    data.Write(BitConverter.GetBytes((byte)data.Length), 0, 1);
+                    // data.Write(BitConverter.GetBytes((byte)data.Length), 0, 1);
+                    data.Write(BitConverter.GetBytes((ushort)data.Length), 0, 1);  // TODO: fix to use byte
                 }
                 res = data.ToArray();
